Alterar tipo dos telefones Visita Integer -> Long

+++ app/src/main/java/hkeller/escolacaesguia/visita/model/Visita.java
@@ -30,13 +30,13 @@ public class Visita {
   private Integer cep;
 
   @Column(name = "telefone_escola")
-  private Integer telefoneEscola;
+  private Long telefoneEscola;
 
   @Column(name = "nome_responsavel", nullable = false)
   private String nomeResponsavel;
 
   @Column(name = "telefone_responsavel")
-  private Integer telefoneResponsavel;
+  private Long telefoneResponsavel;
 
   @Column(name = "numero_alunos")
   private Integer numeroAlunos;
@@ -118,11 +118,11 @@ public class Visita {
     this.cep = cep;
   }
 
-  public Integer getTelefoneEscola() {
+  public Long getTelefoneEscola() {
     return telefoneEscola;
   }
 
-  public void setTelefoneEscola(Integer telefoneEscola) {
+  public void setTelefoneEscola(Long telefoneEscola) {
     this.telefoneEscola = telefoneEscola;
   }
 
@@ -134,11 +134,11 @@ public class Visita {
     this.nomeResponsavel = nomeResponsavel;
   }
 
-  public Integer getTelefoneResponsavel() {
+  public Long getTelefoneResponsavel() {
     return telefoneResponsavel;
   }
 
-  public void setTelefoneResponsavel(Integer telefoneResponsavel) {
+  public void setTelefoneResponsavel(Long telefoneResponsavel) {
     this.telefoneResponsavel = telefoneResponsavel;
   }
